+++ tests/python/frontend/pytorch/test_forward.py
@@ -23,27 +23,15 @@ from tempfile import TemporaryDirectory
 from scipy.stats import t as tdistr
 import numpy as np
 import torch
+from torch.nn import Module
 import tvm
 import torchvision
-import single_op
-
 
 from tvm import relay
 from tvm.contrib import graph_runtime
-#from tvm.relay.testing.config import ctx_list
+from tvm.relay.testing.config import ctx_list
 
 sys.setrecursionlimit(10000)
-
-TARGET = 'llvm'
-CTX = tvm.cpu()
-EXT_ACCEL = None
-
-model_names = []
-baseline_latencies_map = {}
-compiled_latencies_map = {}
-speedups_map = {}
-
-test_repeats = 1
 
 def _vectorize(ten):
     return ten.reshape(-1)
@@ -67,65 +55,26 @@ def assert_shapes_match(tru, est):
         msg = "Output shapes {} and {} don't match"
         raise AssertionError(msg.format(tru.shape, est.shape))
 
-def load_single_op(model_name, input_type=None):
-    """Given a model name, returns a single-operator model in eval
-    mode as well as an example input."""
-    input_shape = [1, 3, 224, 224]
-    if input_type is None or input_type == 'float32':
-        model = getattr(single_op, model_name)().float().eval()
-        input_data = torch.rand(input_shape).float()
-    elif input_type == 'float64':
-        model = getattr(single_op, model_name)().double().eval()
-        #input_data = torch.rand(input_shape).double()
-        #input_data = torch.from_numpy(np.random.random_sample((1, 3, 224, 224))).double()
-        temp = np.random.random_sample((1, 3, 224, 224))
-        input_data = torch.from_numpy(temp)
-    elif input_type == 'float16':
-        model = getattr(single_op, model_name)().half().eval()
-        input_data = torch.rand(input_shape).half()
-    elif input_type == 'int32':
-        model = getattr(single_op, model_name)().eval()
-        input_data = torch.randint(0, 10, input_shape).int()
-    elif input_type == 'int16':
-        model = getattr(single_op, model_name)().eval()
-        input_data = torch.randint(0, 10, input_shape).short()
-    elif input_type == 'int8':
-        model = getattr(single_op, model_name)().eval()
-        input_data = torch.randint(0, 10, input_shape).char()
-    elif input_type == 'uint8':
-        model = getattr(single_op, model_name)().eval()
-        input_data = torch.randint(0, 10, input_shape).byte()
-    return model, input_data
-
-def load_torchvision(model_name, input_type=None):
+def load_torchvision(model_name):
     """Given a model name, returns a Torchvision model in eval mode as well
     as an example input."""
-
-    if model_name.startswith('inception'):
-        height = width = 299
-        mean = [0.5, 0.5, 0.5]
-        std = [0.5, 0.5, 0.5]
-    else:
-        height = width = 224
-        mean = [0.485, 0.456, 0.406]
-        std = [0.229, 0.224, 0.225]
-    input_shape = [1, 3, height, width]
-
-    model = getattr(torchvision.models, model_name)(pretrained=True)
-    #model = model.float().eval()
-
-    if input_type is None or input_type == 'float32':
-        model = model.float().eval()
+    with torch.no_grad():
+        if model_name.startswith('inception'):
+            height = width = 299
+            mean = [0.5, 0.5, 0.5]
+            std = [0.5, 0.5, 0.5]
+        else:
+            height = width = 224
+            mean = [0.485, 0.456, 0.406]
+            std = [0.229, 0.224, 0.225]
+        input_shape = [1, 3, height, width]
         input_data = torch.randn(input_shape).float()
-    elif input_type == 'float64':
-        model = model.double().eval()
-        input_data = torch.randn(input_shape).double()
-
-    for channel in range(3):
-        input_data[:, channel] -= mean[channel]
-        input_data[:, channel] /= std[channel]
-
-    return model, input_data
+        for channel in range(3):
+            input_data[:, channel] -= mean[channel]
+            input_data[:, channel] /= std[channel]
+        model = getattr(torchvision.models, model_name)(pretrained=True)
+        model = model.float().eval()
+        return model, input_data
 
 def load_pretrainedmodels(model_name):
     """Given a model name, returns a pretrainedmodels.pytorch model in eval
@@ -139,10 +88,8 @@ def load_pretrainedmodels(model_name):
         input_data[:, channel] /= model.std[channel]
     return model, input_data
 
-def load_model(model_name, input_type=None):
+def load_model(model_name):
     """Given a model name, returns a model as well as an example input."""
-    if hasattr(single_op, model_name):
-        return load_single_op(model_name, input_type)
     if hasattr(torchvision.models, model_name):
         return load_torchvision(model_name)
     try:
@@ -201,505 +148,725 @@ def measure_latency(model, input_shapes, output_shapes, thresh, dryruns=40):
             if err < thresh:
                 return est
 
-def verify_model(model_name, input_type=None):
+def verify_model(model_name, input_data=[]):
     """Assert that the output of a compiled model matches with that of its
     baseline."""
-    baseline_model, baseline_input = load_model(model_name, input_type)
+    if len(input_data) == 0:
+        baseline_model, baseline_input = load_model(model_name)
+    else:
+        baseline_model = model_name
+        baseline_input = input_data
     if torch.cuda.is_available():
         baseline_model = baseline_model.cuda()
         baseline_input = baseline_input.cuda()
     baseline_outputs = baseline_model(baseline_input)
-    dtype = input_type
-    if input_type is None or input_type == 'float32':
-        baseline_model = baseline_model.float()
-        baseline_input = baseline_input.float()
-        baseline_outputs = baseline_outputs.float()
-        if isinstance(baseline_outputs, tuple):
-            baseline_outputs = tuple(out.detach().cpu().numpy() for out in baseline_outputs)
-        else:
-            baseline_outputs = (baseline_outputs.detach().float().cpu().numpy(),)
-        input_type = 'float32'
-        dtype = 'float32'
-    elif input_type == 'float64':
-        baseline_model = baseline_model.double()
-        baseline_input = baseline_input.double()
-        baseline_outputs = baseline_outputs.double()
-        if isinstance(baseline_outputs, tuple):
-            baseline_outputs = tuple(out.detach().double().cpu().numpy() for out in baseline_outputs)
-        else:
-            baseline_outputs = (baseline_outputs.detach().double().cpu().numpy(),)
-    elif input_type == 'float16':
-        baseline_model = baseline_model.half()
-        baseline_input = baseline_input.half()
-        baseline_outputs = baseline_outputs.half()
-        if isinstance(baseline_outputs, tuple):
-            baseline_outputs = tuple(out.detach().half().cpu().numpy() for out in baseline_outputs)
-        else:
-            baseline_outputs = (baseline_outputs.detach().half().cpu().numpy(),)
-    elif input_type == 'int32':
-        baseline_input = baseline_input.int()
-        baseline_outputs = baseline_outputs.int()
-        if isinstance(baseline_outputs, tuple):
-            baseline_outputs = tuple(out.detach().cpu().numpy() for out in baseline_outputs)
-        else:
-            baseline_outputs = (baseline_outputs.detach().int().cpu().numpy(),)
-    elif input_type == 'int16':
-        baseline_input = baseline_input.short()
-        baseline_outputs = baseline_outputs.short()
-        if isinstance(baseline_outputs, tuple):
-            baseline_outputs = tuple(out.detach().cpu().numpy() for out in baseline_outputs)
-        else:
-            baseline_outputs = (baseline_outputs.detach().short().cpu().numpy(),)
-    elif input_type == 'int8':
-        baseline_input = baseline_input.char()
-        baseline_outputs = baseline_outputs.char()
-        if isinstance(baseline_outputs, tuple):
-            baseline_outputs = tuple(out.detach().cpu().numpy() for out in baseline_outputs)
-        else:
-            baseline_outputs = (baseline_outputs.detach().char().cpu().numpy(),)
-    elif input_type == 'uint8':
-        baseline_input = baseline_input.byte()
-        baseline_outputs = baseline_outputs.byte()
-        if isinstance(baseline_outputs, tuple):
-            baseline_outputs = tuple(out.detach().cpu().numpy() for out in baseline_outputs)
-        else:
-            baseline_outputs = (baseline_outputs.detach().byte().cpu().numpy(),)
+    if isinstance(baseline_outputs, tuple):
+        baseline_outputs = tuple(out.detach().cpu().numpy() for out in baseline_outputs)
+    else:
+        baseline_outputs = (baseline_outputs.detach().float().cpu().numpy(),)
     output_shapes = [out.shape for out in baseline_outputs]
+    dtype = 'float32'
     input_name = 'input0'
     input_shapes = {input_name: list(baseline_input.shape)}
-
-    trace = torch.jit.trace(baseline_model, baseline_input)
-    if input_type is None or input_type == 'float32':
-        trace = trace.float().eval()
-    elif input_type == 'float64':
-        trace = trace.double().eval()
-    elif input_type == 'float16':
-        trace = trace.half().eval()
-    elif input_type == 'int32':
-        trace = trace.float().eval()
-    elif input_type == 'int16':
-        trace = trace.float().eval()
-    elif input_type == 'int8':
-        trace = trace.eval()
-    elif input_type == 'uint8':
-        trace = trace.eval()
+    baseline_model(baseline_input)
+    trace = torch.jit.trace(baseline_model, baseline_input).float().eval()
     if torch.cuda.is_available():
         trace = trace.cuda()
     else:
         trace = trace.cpu()
-    with TemporaryDirectory() as tmp:
-        path = os.path.join(tmp, 'model.pth')
-        torch.jit.save(trace, path)
 
-        mod, params = relay.frontend.from_pytorch(trace, input_shapes)
-
+    mod, params = relay.frontend.from_pytorch(trace, input_shapes)
     compiled_input = {input_name: tvm.nd.array(baseline_input.cpu().numpy())}
 
     with relay.build_config(opt_level=3):
-        relay_graph, relay_lib, relay_params = relay.build(mod, target=TARGET, params=params)
-        relay_model = graph_runtime.create(relay_graph, relay_lib, CTX)
-        relay_model.set_input(**relay_params)
-        relay_model.set_input(**compiled_input)
-        relay_model.run()
-    for i, baseline_output in enumerate(baseline_outputs):
-        output_shape = baseline_output.shape
-        compiled_output = relay_model.get_output(
-            i, tvm.nd.array(np.zeros(output_shape).astype(dtype), CTX)).asnumpy()
-
-        compiled_relay_output = relay_model.get_output(
-            i, tvm.nd.array(np.zeros(output_shape).astype(dtype), CTX)).asnumpy()
-
-        assert_shapes_match(baseline_output, compiled_output)
-        tvm.testing.assert_allclose(baseline_output, compiled_output,
-                                    rtol=1e-3, atol=1e-3)
-
-        assert_shapes_match(baseline_output, compiled_relay_output)
-        tvm.testing.assert_allclose(baseline_output, compiled_relay_output,
-                                    rtol=1e-3, atol=1e-3)
-
-    from subprocess import call
-    call('rm -rf ~/.torch/models/*', shell=True)
-
-def print_results():
-    print(baseline_latencies_map)
-    print(compiled_latencies_map)
-    print(speedups_map)
-
-    thresh = 1e-2
-    units = 1e3
-    thresh = int(thresh * units)
-
-    for model_name in model_names:
-
-        compiled_sum = 0.0
-        baseline_sum = 0.0
-        speedup_sum = 0.0
-
-        print("For model name "+model_name)
-        for i in range(0, test_repeats):
-            print(f'Compiled latency is {compiled_latencies_map[model_name][i]:.3f} +/- {thresh:d} ms.')
-            print(f'Baseline latency is {baseline_latencies_map[model_name][i]:.3f} +/- {thresh:d} ms.')
-            print(f'Relative speedup is {speedups_map[model_name][i]:.3f}')
-
-            compiled_sum = compiled_sum + compiled_latencies_map[model_name][i]
-            baseline_sum = baseline_sum + baseline_latencies_map[model_name][i]
-            speedup_sum = speedup_sum + speedups_map[model_name][i]
-
-        print(f'Average compiled latency is {compiled_sum/test_repeats:.3f} +/- {thresh:d} ms.')
-        print(f'Average baseline latency is {baseline_sum/test_repeats:.3f} +/- {thresh:d} ms.')
-        print(f'Average relative speedup is {speedup_sum/test_repeats:.3f}')
-
-"""
-# Test Functions
-def test_add1():
-    verify_model('Add1')
-
-def test_add1int16():
-    verify_model('Add1', input_type='int16')
-
-def test_add2():
-    verify_model('Add2')
-
-def test_add3():
-    verify_model('Add3')
-
-def test_add4():
-    verify_model('Add4')
-
-def test_add5():
-    verify_model('Add5')
-
-def test_add3int32():
-    verify_model('Add3Int32', input_type='int32')
-
-def test_add4int32():
-    verify_model('Add4Int32', input_type='int32')
-
-
-def test_add3float16():
-    verify_model('Add3Float16', input_type='float16')
-
-def test_add4float16():
-    verify_model('Add4Float16', input_type='float16')
-
-def test_add3float64():
-    verify_model('Add3Float64', input_type='float64')
-
-def test_add4float64():
-    verify_model('Add4Float64', input_type='float64')
-
-def test_subtract1():
-    verify_model('Subtract1')
-
-def test_subtract1int32():
-    verify_model('Subtract1', input_type='int32')
-
-def test_subtract1int16():
-    verify_model('Subtract1', input_type='int16')
-
-def test_subtract1int8():
-    verify_model('Subtract1', input_type='int8')
-
-def test_subtract1uint8():
-    verify_model('Subtract1', input_type='uint8')
-
-def test_subtract2():
-    verify_model('Subtract2')
-
-def test_subtract3():
-    verify_model('Subtract3')
-
-def test_subtract3int32():
-    verify_model('Subtract3Int32', input_type='int32')
-
-def test_subtract4():
-    verify_model('Subtract4')
-
-def test_subtract5():
-    verify_model('Subtract5')
-
-def test_multiply1():
-    verify_model('Multiply1')
-
-def test_multiply2():
-    verify_model('Multiply2')
-
-def test_multiply3():
-    verify_model('Multiply3')
-
-def test_multiply4():
-    verify_model('Multiply4')
-
-def test_multiply5():
-    verify_model('Multiply5')
-
-def test_unsqueeze1():
-    verify_model('Unsqueeze1')
-
-def test_concatenate1():
-    verify_model('Concatenate1')
-
-def test_concatenate1():
-    verify_model('Concatenate1')
-
-def test_concatenate2():
-    verify_model('Concatenate2')
-
-def test_relu1():
-    verify_model('ReLU1')
-
-def test_adaptiveavgpool2d1():
-    verify_model('AdaptiveAvgPool2D1')
-
-def test_adaptiveavgpool2d2():
-    verify_model('AdaptiveAvgPool2D2')
-
-def test_adaptiveavgpool2d3():
-    verify_model('AdaptiveAvgPool2D3')
-
-def test_maxpool2d1():
-    verify_model('MaxPool2D1')
-
-def test_maxpool2d2():
-    verify_model('MaxPool2D2')
-
-def test_maxpool2d3():
-    verify_model('MaxPool2D3')
-
-def test_hardtanh1():
-    verify_model('HardTanh1')
-
-def test_conv2d1():
-    verify_model('Conv2D1')
-
-def test_conv2d1float64():
-    verify_model('Conv2D1Float64', input_type='float64')
-
-def test_conv2d2():
-    verify_model('Conv2D2')
-
-def test_threshold1():
-    verify_model('Threshold1')
-
-def test_contiguous1():
-    verify_model('Contiguous1')
-
-def test_batchnorm1():
-    verify_model('BatchNorm1', input_type='float32')
-
-def test_batchnorm1float64():
-    verify_model('BatchNorm1Float64', input_type='float64')
-
-def test_batchnorm2():
-    verify_model('BatchNorm2')
-
-def test_transpose1():
-    verify_model('Transpose1')
-
-def test_transpose2():
-    verify_model('Transpose2')
-
-def test_size1():
-    verify_model('Size1')
-
-def test_view1():
-    verify_model('View1')
-
-def test_view2():
-    verify_model('View2')
-
-def test_select1():
-    verify_model('Select1')
-
-def test_clone1():
-    verify_model('Clone1')
-
-def test_logsoftmax1():
-    verify_model('LogSoftmax1')
-
-def test_sigmoid1():
-    verify_model('Sigmoid1')
-
-def test_dense1():
-    verify_model('Dense1')
-
-def test_dense2():
-    verify_model('Dense2')
-
-def test_avgpool2d1():
-    verify_model('AvgPool2D1')
-
-def test_dropout1():
-    verify_model('Dropout1')
-
-def test_slice1():
-    verify_model('Slice1')
-
-def test_slice2():
-    verify_model('Slice2')
-
-def test_mean1():
-    verify_model('Mean1')
-
-def test_expand1():
-    verify_model('Expand1')
-
-def test_pow1():
-    verify_model('Pow1')
-
-def test_chunk1():
-    verify_model('Chunk1')
+        for target, ctx in ctx_list():
+            relay_graph, relay_lib, relay_params = relay.build(mod, target=target, params=params)
+            relay_model = graph_runtime.create(relay_graph, relay_lib, ctx)
+            relay_model.set_input(**relay_params)
+            relay_model.set_input(**compiled_input)
+            relay_model.run()
+
+            for i, baseline_output in enumerate(baseline_outputs):
+                output_shape = baseline_output.shape
+                compiled_output = relay_model.get_output(
+                    i, tvm.nd.array(np.zeros(output_shape).astype(dtype), ctx)).asnumpy()
+
+                compiled_relay_output = relay_model.get_output(
+                    i, tvm.nd.array(np.zeros(output_shape).astype(dtype), ctx)).asnumpy()
+
+                assert_shapes_match(baseline_output, compiled_output)
+                tvm.testing.assert_allclose(baseline_output, compiled_output,
+                                            rtol=1e-3, atol=1e-3)
+
+                assert_shapes_match(baseline_output, compiled_relay_output)
+                tvm.testing.assert_allclose(baseline_output, compiled_relay_output,
+                                            rtol=1e-3, atol=1e-3)
+
+    # Try manually removing model from memory after each test
+
+    if torch.cuda.is_available():
+        # Print memory usage
+        print(torch.cuda.memory_allocated(0))
+        print(torch.cuda.max_memory_allocated(0))
+
+    del model_name
+    del baseline_model
+    torch.cuda.empty_cache()
+    cpuStats()
+    memReport()
+
+# Memory checking
+def memReport():
+    import gc
+    for obj in gc.get_objects():
+        if torch.is_tensor(obj):
+            print(type(obj), obj.size())
+
+def cpuStats():
+    import psutil
+    print(sys.version)
+    print(psutil.cpu_percent())
+    print(psutil.virtual_memory())  # physical memory usage
+    pid = os.getpid()
+    py = psutil.Process(pid)
+    memoryUse = py.memory_info()[0] / 2. ** 30  # memory use in GB...I think
+    print('memory GB:', memoryUse)
+
+# Single operator tests
+def test_forward_add():
+    torch.set_grad_enabled(False)
+    input_shape = [10]
+
+    class Add1(Module):
+        def forward(self, *args):
+            return args[0] + args[0]
+
+    class Add2(Module):
+        def forward(self, *args):
+            return args[0] + 1
+
+    class Add3(Module):
+        def forward(self, *args):
+            ones = torch.ones(input_shape, dtype=torch.float)
+            if torch.cuda.is_available():
+                ones = ones.cuda()
+            return args[0] + ones
+
+    class Add4(Module):
+        def forward(self, *args):
+            ones = torch.ones(input_shape, dtype=torch.float)
+            if torch.cuda.is_available():
+                ones = ones.cuda()
+            return args[0] + ones
+
+    class Add5(Module):
+        def forward(self, *args):
+            ones = torch.ones([], dtype=torch.float)
+            if torch.cuda.is_available():
+                ones = ones.cuda()
+            return args[0] + ones
+
+    with torch.no_grad():
+        input_data = torch.rand(input_shape).float()
+        verify_model(Add1().float().eval(), input_data=input_data)
+        verify_model(Add2().float().eval(), input_data=input_data)
+        verify_model(Add3().float().eval(), input_data=input_data)
+        verify_model(Add4().float().eval(), input_data=input_data)
+        verify_model(Add5().float().eval(), input_data=input_data)
+
+def test_forward_subtract():
+    torch.set_grad_enabled(False)
+    input_shape = [10]
+
+    class Subtract1(Module):
+        def forward(self, *args):
+            return args[0] - args[0]
+
+    class Subtract2(Module):
+        def forward(self, *args):
+            return args[0] - 1
+
+    class Subtract3(Module):
+        def forward(self, *args):
+            ones = torch.ones(input_shape)
+            if torch.cuda.is_available():
+                ones = ones.cuda()
+            return args[0] - ones
+
+    class Subtract4(Module):
+        def forward(self, *args):
+            ones = torch.ones(input_shape)
+            if torch.cuda.is_available():
+                ones = ones.cuda()
+            return args[0] - ones
+
+    class Subtract5(Module):
+        def forward(self, *args):
+            ones = torch.ones([])
+            if torch.cuda.is_available():
+                ones = ones.cuda()
+            return args[0] - ones
+
+    with torch.no_grad():
+        input_data = torch.rand(input_shape).float()
+        verify_model(Subtract1().float().eval(), input_data=input_data)
+        verify_model(Subtract2().float().eval(), input_data=input_data)
+        verify_model(Subtract3().float().eval(), input_data=input_data)
+        verify_model(Subtract4().float().eval(), input_data=input_data)
+        verify_model(Subtract5().float().eval(), input_data=input_data)
+
+def test_forward_multiply():
+    torch.set_grad_enabled(False)
+    input_shape = [10]
+
+    class Multiply1(Module):
+        def forward(self, *args):
+            return args[0] * args[0]
+
+    class Multiply2(Module):
+        def forward(self, *args):
+            return args[0] * 1
+
+    class Multiply3(Module):
+        def forward(self, *args):
+            ones = torch.ones(input_shape)
+            if torch.cuda.is_available():
+                ones = ones.cuda()
+            return args[0] * ones
+
+    class Multiply4(Module):
+        def forward(self, *args):
+            ones = torch.ones(input_shape)
+            if torch.cuda.is_available():
+                ones = ones.cuda()
+            return args[0] * ones
+
+    class Multiply5(Module):
+        def forward(self, *args):
+            ones = torch.ones([])
+            if torch.cuda.is_available():
+                ones = ones.cuda()
+            return args[0] * ones
+
+    with torch.no_grad():
+        input_data = torch.rand(input_shape).float()
+        verify_model(Multiply1().float().eval(), input_data=input_data)
+        verify_model(Multiply2().float().eval(), input_data=input_data)
+        verify_model(Multiply3().float().eval(), input_data=input_data)
+        verify_model(Multiply4().float().eval(), input_data=input_data)
+        verify_model(Multiply5().float().eval(), input_data=input_data)
+
+def test_forward_unsqueeze():
+    torch.set_grad_enabled(False)
+    input_shape = [10, 10]
+
+    class Unsqueeze1(Module):
+        def forward(self, *args):
+            return args[0].unsqueeze(2)
+
+    input_data = torch.rand(input_shape).float()
+    verify_model(Unsqueeze1().float().eval(), input_data=input_data)
+
+def test_forward_concatenate():
+    torch.set_grad_enabled(False)
+    input_shape = [1, 3, 10, 10]
+
+    class Concatenate1(Module):
+        def forward(self, *args):
+            return torch.cat([args[0][:, 0].unsqueeze(1), args[0][:, 1].unsqueeze(1)], 1)
+
+    class Concatenate2(Module):
+        def forward(self, *args):
+            a = (args[0][:, :, 0] + 2) * 7
+            b = (args[0][:, :, 1] + 3) * 11
+            c = (args[0][:, :, 2] + 5) * 13
+            return torch.cat([t.unsqueeze(2) for t in [a, b, c]], 2)
+
+    with torch.no_grad():
+        input_data = torch.rand(input_shape).float()
+        verify_model(Concatenate1().float().eval(), input_data=input_data)
+        verify_model(Concatenate2().float().eval(), input_data=input_data)
+
+def test_forward_relu():
+    torch.set_grad_enabled(False)
+    input_shape = [10, 10]
+
+    class ReLU1(Module):
+        def forward(self, *args):
+            return torch.nn.ReLU()(args[0])
+
+    with torch.no_grad():
+        input_data = torch.rand(input_shape).float()
+        verify_model(ReLU1().float().eval(), input_data=input_data)
+
+def test_forward_adaptiveavgpool():
+    torch.set_grad_enabled(False)
+    input_shape = [1, 3, 224, 224]
+
+    class AdaptiveAvgPool2D1(Module):
+        def forward(self, *args):
+            return torch.nn.AdaptiveAvgPool2d([1, 1])(args[0])
+
+    class AdaptiveAvgPool2D2(Module):
+        def forward(self, *args):
+            return torch.nn.AdaptiveAvgPool2d([100, 100])(args[0])
+
+    class AdaptiveAvgPool2D3(Module):
+        def forward(self, *args):
+            return torch.nn.AdaptiveAvgPool2d([224, 224])(args[0])
+
+    with torch.no_grad():
+        input_data = torch.rand(input_shape).float()
+        verify_model(AdaptiveAvgPool2D1().float().eval(), input_data=input_data)
+        verify_model(AdaptiveAvgPool2D2().float().eval(), input_data=input_data)
+        verify_model(AdaptiveAvgPool2D3().float().eval(), input_data=input_data)
+
+def test_forward_maxpool():
+    torch.set_grad_enabled(False)
+    input_shape = [1, 3, 224, 224]
+
+    class MaxPool2D1(Module):
+        def forward(self, *args):
+            return torch.nn.MaxPool2d(kernel_size=[1, 1])(args[0])
+
+    class MaxPool2D2(Module):
+        def forward(self, *args):
+            return torch.nn.MaxPool2d(kernel_size=[100, 100])(args[0])
+
+    class MaxPool2D3(Module):
+        def forward(self, *args):
+            return torch.nn.MaxPool2d(kernel_size=[224, 224])(args[0])
+
+    with torch.no_grad():
+        input_data = torch.rand(input_shape).float()
+        verify_model(MaxPool2D1().float().eval(), input_data=input_data)
+        verify_model(MaxPool2D2().float().eval(), input_data=input_data)
+        verify_model(MaxPool2D3().float().eval(), input_data=input_data)
+
+def test_forward_avgpool():
+    torch.set_grad_enabled(False)
+    input_shape = [1, 3, 100, 100]
+
+    class AvgPool2D1(Module):
+        def forward(self, *args):
+            return torch.nn.AvgPool2d(kernel_size=[100, 100])(args[0])
+
+    with torch.no_grad():
+        input_data = torch.rand(input_shape).float()
+        verify_model(AvgPool2D1().float().eval(), input_data=input_data)
+
+def test_forward_hardtanh():
+    torch.set_grad_enabled(False)
+    input_shape = [10]
+
+    class HardTanh1(Module):
+        def forward(self, *args):
+            return torch.nn.Hardtanh()(args[0])
+
+    with torch.no_grad():
+        input_data = torch.rand(input_shape).float()
+        verify_model(HardTanh1().float().eval(), input_data=input_data)
+
+def test_forward_conv():
+    torch.set_grad_enabled(False)
+    input_shape = [1, 3, 100, 100]
+
+    class Conv2D1(Module):
+        def __init__(self):
+            super(Conv2D1, self).__init__()
+            self.conv = torch.nn.Conv2d(3, 64, 7, bias=True)
+            self.softmax = torch.nn.Softmax()
+
+        def forward(self, *args):
+            return self.softmax(self.conv(args[0]))
+
+    class Conv2D2(Module):
+        def __init__(self):
+            super(Conv2D2, self).__init__()
+            self.conv = torch.nn.Conv2d(3, 64, 7, bias=False)
+            self.softmax = torch.nn.Softmax()
+
+        def forward(self, *args):
+            return self.softmax(self.conv(args[0]))
+
+    with torch.no_grad():
+        input_data = torch.rand(input_shape).float()
+        verify_model(Conv2D1().float().eval(), input_data=input_data)
+        verify_model(Conv2D2().float().eval(), input_data=input_data)
+
+def test_forward_threshold():
+    torch.set_grad_enabled(False)
+    input_shape = [1, 3]
+
+    class Threshold1(Module):
+        def forward(self, *args):
+            return torch.nn.Threshold(0, 0)(args[0])
+
+    with torch.no_grad():
+        input_data = torch.rand(input_shape).float()
+        verify_model(Threshold1().float().eval(), input_data=input_data)
+
+def test_forward_contiguous():
+    torch.set_grad_enabled(False)
+    input_shape = [10]
+
+    class Contiguous1(Module):
+        def forward(self, *args):
+            return args[0].contiguous()
+
+    with torch.no_grad():
+        input_data = torch.rand(input_shape).float()
+        verify_model(Contiguous1().float().eval(), input_data=input_data)
+
+def test_forward_batchnorm():
+    torch.set_grad_enabled(False)
+    input_shape = [1, 3, 10, 10]
+
+    class BatchNorm1(Module):
+        def __init__(self):
+            super(BatchNorm1, self).__init__()
+            self.batch_norm = torch.nn.BatchNorm2d(3, affine=True)
+        def forward(self, *args):
+            return self.batch_norm(args[0])
+
+    class BatchNorm2(Module):
+        def __init__(self):
+            super(BatchNorm2, self).__init__()
+            self.batch_norm = torch.nn.BatchNorm2d(3, affine=False)
+        def forward(self, *args):
+            return self.batch_norm(args[0])
+
+    with torch.no_grad():
+        input_data = torch.rand(input_shape).float()
+        verify_model(BatchNorm1().float().eval(), input_data=input_data)
+        verify_model(BatchNorm2().float().eval(), input_data=input_data)
+
+def test_forward_transpose():
+    torch.set_grad_enabled(False)
+    input_shape = [1, 3, 10, 10]
+
+    class Transpose1(Module):
+        def forward(self, *args):
+            return args[0].transpose(2, 3)
+
+    class Transpose2(Module):
+        def forward(self, *args):
+            return args[0].transpose(-2, -1)
+
+    with torch.no_grad():
+        input_data = torch.rand(input_shape).float()
+        verify_model(Transpose1().float().eval(), input_data=input_data)
+        verify_model(Transpose2().float().eval(), input_data=input_data)
+
+def test_forward_size():
+    torch.set_grad_enabled(False)
+    input_shape = [1, 3]
+
+    class Size1(Module):
+        def forward(self, *args):
+            return args[0].size(0) * args[0]
+
+    with torch.no_grad():
+        input_data = torch.rand(input_shape).float()
+        verify_model(Size1().float().eval(), input_data=input_data)
+
+def test_forward_view():
+    torch.set_grad_enabled(False)
+    input_shape = [1, 3, 224, 224]
+
+    class View1(Module):
+        def forward(self, *args):
+            return args[0].view((1, 3 * 224 * 224))
+
+    class View2(Module):
+        def forward(self, *args):
+            return args[0].view(args[0].shape[0], -1)
+
+    with torch.no_grad():
+        input_data = torch.rand(input_shape).float()
+        verify_model(View1().float().eval(), input_data=input_data)
+        verify_model(View2().float().eval(), input_data=input_data)
+
+def test_forward_select():
+    torch.set_grad_enabled(False)
+    input_shape = [1, 3, 10, 10]
+
+    class Select1(Module):
+        def forward(self, *args):
+            return args[0].select(1, 1)
+
+    with torch.no_grad():
+        input_data = torch.rand(input_shape).float()
+        verify_model(Select1().float().eval(), input_data=input_data)
+
+def test_forward_clone():
+    torch.set_grad_enabled(False)
+    input_shape = [10]
+
+    class Clone1(Module):
+        def forward(self, *args):
+            return args[0].clone()
+
+    with torch.no_grad():
+        input_data = torch.rand(input_shape).float()
+        verify_model(Clone1().float().eval(), input_data=input_data)
+
+def test_forward_logsoftmax():
+    torch.set_grad_enabled(False)
+    input_shape = [1, 3, 10, 10]
+
+    class LogSoftmax1(Module):
+        def forward(self, *args):
+            return torch.nn.LogSoftmax(dim=1)(args[0][0, 0])
+
+    with torch.no_grad():
+        input_data = torch.rand(input_shape).float()
+        verify_model(LogSoftmax1().float().eval(), input_data=input_data)
+
+def test_forward_sigmoid():
+    torch.set_grad_enabled(False)
+    input_shape = [1, 3, 10, 10]
+
+    class Sigmoid1(Module):
+        def forward(self, *args):
+            return torch.nn.Sigmoid()(args[0])
+
+    with torch.no_grad():
+        input_data = torch.rand(input_shape).float()
+        verify_model(Sigmoid1().float().eval(), input_data=input_data)
+
+def test_forward_dense():
+    torch.set_grad_enabled(False)
+    input_shape = [1, 3, 10, 10]
+
+    class Dense1(Module):
+        def __init__(self):
+            super(Dense1, self).__init__()
+            self.linear = torch.nn.Linear(10, 7, bias=True)
+        def forward(self, *args):
+            return self.linear(args[0][0, 0])
+
+    class Dense2(Module):
+        def __init__(self):
+            super(Dense2, self).__init__()
+            self.linear = torch.nn.Linear(10, 7, bias=False)
+        def forward(self, *args):
+            return self.linear(args[0][0, 0])
+
+    with torch.no_grad():
+        input_data = torch.rand(input_shape).float()
+        verify_model(Dense1().float().eval(), input_data=input_data)
+        verify_model(Dense2().float().eval(), input_data=input_data)
+
+def test_forward_dropout():
+    torch.set_grad_enabled(False)
+    input_shape = [1, 3, 10, 10]
+
+    class Dropout1(Module):
+        def forward(self, *args):
+            return torch.nn.functional.dropout(args[0][0, 0], 0.5, False)
+
+    with torch.no_grad():
+        input_data = torch.rand(input_shape).float()
+        verify_model(Dropout1().float().eval(), input_data=input_data)
+
+def test_forward_slice():
+    torch.set_grad_enabled(False)
+    input_shape = [1, 3, 10, 10]
+
+    class Slice1(Module):
+        def forward(self, *args):
+            return args[0][:, :, :, :3]
+
+    class Slice2(Module):
+        def forward(self, *args):
+            return args[0][0, :, :, :]
+
+    with torch.no_grad():
+        input_data = torch.rand(input_shape).float()
+        verify_model(Slice1().float().eval(), input_data=input_data)
+        verify_model(Slice2().float().eval(), input_data=input_data)
+
+def test_forward_mean():
+    torch.set_grad_enabled(False)
+    input_shape = [1, 3, 10, 10]
+
+    class Mean1(Module):
+        def forward(self, *args):
+            return args[0].mean(2)
+
+    with torch.no_grad():
+        input_data = torch.rand(input_shape).float()
+        verify_model(Mean1().float().eval(), input_data=input_data)
+
+def test_forward_expand():
+    torch.set_grad_enabled(False)
+    input_shape = [1, 3, 10, 10]
+
+    class Expand1(Module):
+        def forward(self, *args):
+            return args[0].expand((3, -1, -1, -1))
+
+    with torch.no_grad():
+        input_data = torch.rand(input_shape).float()
+        verify_model(Expand1().float().eval(), input_data=input_data)
+
+def test_forward_pow():
+    torch.set_grad_enabled(False)
+    input_shape = [1, 3, 10, 10]
+
+    class Pow1(Module):
+        def forward(self, *args):
+            return args[0] ** 2
+
+    with torch.no_grad():
+        input_data = torch.rand(input_shape).float()
+        verify_model(Pow1().float().eval(), input_data=input_data)
+
+def test_forward_chunk():
+    torch.set_grad_enabled(False)
+    input_shape = [1, 3, 14, 14]
+
+    class Chunk1(Module):
+        def forward(self, *args):
+            chunks = args[0].chunk(7, 2)
+            return torch.cat(chunks, 2)
+
+    with torch.no_grad():
+        input_data = torch.rand(input_shape).float()
+        verify_model(Chunk1().float().eval(), input_data=input_data)
 
 # Model tests
 def test_resnet18():
+    torch.set_grad_enabled(False)
     verify_model('resnet18')
 
-def test_resnet18float64():
-    verify_model('resnet18', input_type='float64')
-
 def test_resnet34():
+    torch.set_grad_enabled(False)
     verify_model('resnet34')
 
 def test_resnet50():
+    torch.set_grad_enabled(False)
     verify_model('resnet50')
 
 def test_resnet101():
+    torch.set_grad_enabled(False)
     verify_model('resnet101')
 
 def test_resnet152():
+    torch.set_grad_enabled(False)
     verify_model('resnet152')
 
 def test_squeezenet1_0():
+    torch.set_grad_enabled(False)
     verify_model('squeezenet1_0')
 
 def test_squeezenet1_1():
+    torch.set_grad_enabled(False)
     verify_model('squeezenet1_1')
 
-def test_vgg11():
-    verify_model('vgg11')
-
-def test_vgg11float64():
-    verify_model('vgg11', input_type='float64')
-
-def test_vgg13():
-    verify_model('vgg13')
-
-def test_vgg16():
-    verify_model('vgg16')
-
-def test_vgg19():
-    verify_model('vgg19')
-
-def test_vgg11_bn():
-    verify_model('vgg11_bn')
-
-def test_vgg13_bn():
-    verify_model('vgg13_bn')
-
-def test_vgg19_bn():
-    verify_model('vgg19_bn')
-
 def test_mobilenet_v2():
+    torch.set_grad_enabled(False)
     verify_model('mobilenet_v2')
 
 def test_densenet121():
+    torch.set_grad_enabled(False)
     verify_model('densenet121')
 
 def test_densenet161():
+    torch.set_grad_enabled(False)
     verify_model('densenet161')
 
 def test_densenet169():
+    torch.set_grad_enabled(False)
     verify_model('densenet169')
 
 def test_densenet201():
+    torch.set_grad_enabled(False)
     verify_model('densenet201')
 
 def test_inception_v3():
+    torch.set_grad_enabled(False)
     verify_model('inception_v3')
 
-def test_alexnet():
-    verify_model('alexnet')
-
-def test_alexnetfloat64():
-    verify_model('alexnet', input_type='float64')
-
 def test_googlenet():
+    torch.set_grad_enabled(False)
     verify_model('googlenet')
 
 def test_mnasnet0_5():
+    torch.set_grad_enabled(False)
     verify_model('mnasnet0_5')
 
 def test_mnasnet1_0():
+    torch.set_grad_enabled(False)
     verify_model('mnasnet1_0')
+
+#TODO: Fix VGG and AlexNet issues (probably due to pooling)
+"""
+def test_alexnet():
+    torch.set_grad_enabled(False)
+    verify_model('alexnet')
+
+def test_vgg11():
+    torch.set_grad_enabled(False)
+    verify_model('vgg11')
+
+def test_vgg13():
+    torch.set_grad_enabled(False)
+    verify_model('vgg13')
+
+def test_vgg16():
+    torch.set_grad_enabled(False)
+    verify_model('vgg16')
+
+def test_vgg19():
+    torch.set_grad_enabled(False)
+    verify_model('vgg19')
+
+def test_vgg11_bn():
+    torch.set_grad_enabled(False)
+    verify_model('vgg11_bn')
+
+def test_vgg13_bn():
+    torch.set_grad_enabled(False)
+    verify_model('vgg13_bn')
+
+def test_vgg19_bn():
+    torch.set_grad_enabled(False)
+    verify_model('vgg19_bn')
 """
 
 if __name__ == '__main__':
 
-    # TODO: Refactor how testing works for different types
-    test_add3float64()
-    test_add4int32()
-    test_batchnorm1float64()
-    test_subtract3int32()
-    test_subtract1int32()
-    test_subtract1int16()
-    test_subtract1int8()
-    test_subtract1uint8()
-    test_conv2d1float64()
-    test_alexnetfloat64()
-    test_resnet18float64()
-
     # Single operator tests
-    test_add1()
-    test_add2()
-    test_add3()
-    test_add4()
-    test_add5()
-    test_subtract1()
-    test_subtract2()
-    test_subtract3()
-    test_subtract4()
-    test_subtract5()
-    test_multiply1()
-    test_multiply2()
-    test_multiply3()
-    test_multiply4()
-    test_multiply5()
-    test_unsqueeze1()
-    test_concatenate1()
-    test_concatenate2()
-    test_relu1()
-    test_adaptiveavgpool2d1()
-    test_adaptiveavgpool2d2()
-    test_adaptiveavgpool2d3()
-    test_maxpool2d1()
-    test_maxpool2d2()
-    test_maxpool2d3()
-    test_hardtanh1()
-    test_conv2d1()
-    test_conv2d2()
-    test_threshold1()
-    test_contiguous1()
-    test_batchnorm1()
-    test_batchnorm2()
-    test_transpose1()
-    test_transpose2()
-    test_size1()
-    test_view1()
-    test_view2()
-    test_select1()
-    test_clone1()
-    test_logsoftmax1()
-    test_sigmoid1()
-    test_dense1()
-    test_dense2()
-    test_avgpool2d1()
-    test_dropout1()
-    test_slice1()
-    test_slice2()
-    test_mean1()
-    test_expand1()
-    test_pow1()
-    test_chunk1()
+    test_forward_add()
+    test_forward_subtract()
+    test_forward_multiply()
+    test_forward_unsqueeze()
+    test_forward_concatenate()
+    test_forward_relu()
+    test_forward_adaptiveavgpool()
+    test_forward_maxpool()
+    test_forward_hardtanh()
+    test_forward_conv()
+    test_forward_threshold()
+    test_forward_contiguous()
+    test_forward_batchnorm()
+    test_forward_transpose()
+    test_forward_size()
+    test_forward_view()
+    test_forward_select()
+    test_forward_clone()
+    test_forward_logsoftmax()
+    test_forward_sigmoid()
+    test_forward_dense()
+    test_forward_avgpool()
+    test_forward_dropout()
+    test_forward_slice()
+    test_forward_mean()
+    test_forward_expand()
+    test_forward_pow()
+    test_forward_chunk()
 
     # Model tests
     test_resnet18()
@@ -709,20 +876,12 @@ if __name__ == '__main__':
     test_resnet152()
     test_squeezenet1_0()
     test_squeezenet1_1()
-    test_vgg11()
-    test_vgg13()
-    test_vgg16()
-    test_vgg19()
-    test_vgg11_bn()
-    test_vgg13_bn()
-    test_vgg19_bn()
     test_mobilenet_v2()
     test_densenet121()
     test_densenet161()
     test_densenet169()
     test_densenet201()
     test_inception_v3()
-    test_alexnet()
     test_googlenet()
     test_mnasnet0_5()
     test_mnasnet1_0()
